Add initial multi-file version registry documents
Since each of the services already have a version discovery document which
is very suitable, except for containing some formatting differences and some
additional data, define a format that is mechanically derivable from the
existing documents.

It should be noted that a mapping must exist _somewhere_ between project
names and service types. This does not attempt to solve that, but the
process described within assumes the existence of the
service-types-registry.

It should be also noted that this list is not complete - it is merely
the list of available services I could check by hand on Vexxhost, which
I happen to know is running Newton. Hopefully it is both enough of a
straw-man document and a complete enough process for how projects might
maintain such a document going forward that we can finish it as a group
exercise.

Change-Id: If50f6b0ef39cd37ac8640bc653e89d98eb4963b3