Add API version history for Nova
Going back to essex this is fairly easy since we had
API samples for the / versions document in the integrated
tests.

After that it gets a bit weird.

We have diablo-eol in github and the versions can be pulled
from code and unit tests there.

We don't have -eol tags for austin, bexar or cactus, so
we have to check tags.

The last release for cactus was 2011.2 and we can pull
the versions from code and tests.

The last release for bexar was 2011.3.1 and we can pull
the versions from code and tests.

We don't have any release tag for austin in the git repo
even though the openstack releases page says 2010.1 was
the austin release. We can find the 2010.1 nova code on
launchpad:

https://launchpad.net/nova/austin/2010.1

And looking in nova/api/__init__.py we see the
osapi_versions method on the API router class, which shows
the same version (v1.0) as in Bexar.

Per the validation in this repo, all 'links' and 'updated'
keys in responses are removed even if they are in the
actual responses.

This change also updates the schema, and fixes part of it:

1. The README says status can include "EXPERIMENTAL" but that
   wasn't defined as an option in the schema.
2. The "required" key was misspelled.
3. Nova uses "version" instead of "max_version" for microversions,
   so that is added to the schema and README.
4. And for crying in the soup Martha, format the json schema!

Change-Id: I4f353451f2cd2e06c1bb34d695f926e88108c219